moved math conversion function
